requirements.txt
wheel
ordered-set
nuitka
numpy
scipy
pywavelets
pillow
pg8000
pymysql
